+++ cea/interfaces/dashboard/map_layers/solar_irradiance/layers.py
@@ -1,13 +1,13 @@
 import itertools
 import os
 from concurrent.futures import ThreadPoolExecutor
-from typing import Tuple
 
 import fiona
 import pandas as pd
 from pyproj import CRS, Transformer
 
 from cea.inputlocator import InputLocator
+from cea.interfaces.dashboard.map_layers import day_range_to_hour_range
 from cea.interfaces.dashboard.map_layers.base import MapLayer
 from cea.interfaces.dashboard.map_layers.solar_irradiance import SolarIrradiationCategory
 
@@ -137,15 +137,3 @@ class SolarIrradiationMapLayer(MapLayer):
             }
         }
         return output
-
-def day_range_to_hour_range(nth_day_start: int, nth_day_end: int) -> Tuple[int, int]:
-    """
-    Converts a nth day range (e.g. 01-Jan is 1, 31-Dec is 365) to hour range (zero-indexed),
-    where the first hour is the hour at the start of the "start day" and the second hour is the
-    hour at the end of the "end day".
-
-    e.g. day_range_to_hour_range(1, 1) returns 0, 23
-    e.g. day_range_to_hour_range(1, 2) returns 0, 47
-    e.g. day_range_to_hour_range(365, 1) returns 8736, 23
-    """
-    return (nth_day_start-1) * 24, nth_day_end * 24 -1
